comment ajax 작성
중복 작업으로 일시 중단, 장한얼 조원의 작업과 상호 보완예정

+++ src/main/java/com/fashionlog/controller/CommentController.java
@@ -43,17 +43,24 @@ public class CommentController {
 	}
 
 	@RequestMapping("/insertComment")
+	@ResponseBody
 	public String insertComment(Comment comment, @AuthenticationPrincipal SecurityUser securityUser) {
 		Member user = securityUser.getMember(); 
 		comment.setMemberNo(user);
 		commentRepository.save(comment);
-		return "redirect:/comment";
+		return "abab";
 	}
 
 	@RequestMapping("/deleteComment")
-	public String deleteComment(@RequestParam("commentNo")int commentNo) {
-		commentRepository.deleteById(commentNo);
-		return "redirect:/comment";
+	@ResponseBody
+	public String deleteComment(int commentNo, @AuthenticationPrincipal SecurityUser securityUser) {
+		Member user = securityUser.getMember();
+		Comment comm = commentRepository.findById(commentNo).get();
+		if (comm.getMemberNo().getId() == user.getId()){
+			commentRepository.deleteById(commentNo);
+			return commentNo+" is deleted";
+		}
+		return "delete denied";
 	}
 
 }
